+++ supvisors/supervisordata.py
@@ -109,7 +109,7 @@ class SupervisorData(object):
         """
         if not self._system_rpc_interface:
             # the first handler is the XML-RPC interface for rapid access
-            handler = self.httpserver.handlers[0]
+            handler = self.http_server.handlers[0]
             # if authentication used, handler is wrapped
             if self.username:
                 handler = handler.handler
@@ -125,7 +125,7 @@ class SupervisorData(object):
         """
         if not self._supervisor_rpc_interface:
             # the first handler is the XML-RPC interface for rapid access
-            handler = self.httpserver.handlers[0]
+            handler = self.http_server.handlers[0]
             # if authentication used, handler is wrapped
             if self.username:
                 handler = handler.handler
@@ -141,7 +141,7 @@ class SupervisorData(object):
         """
         if not self._supvisors_rpc_interface:
             # the first handler is the XML-RPC interface for rapid access
-            handler = self.httpserver.handlers[0]
+            handler = self.http_server.handlers[0]
             # if authentication is used, handler is wrapped
             if self.username:
                 handler = handler.handler
@@ -157,7 +157,7 @@ class SupervisorData(object):
         return self.supervisord.options.identifier
 
     @property
-    def httpserver(self):
+    def http_server(self):
         """ Get the internal Supervisor HTTP server structure.
 
         :return: the HTTP server structure
@@ -167,11 +167,11 @@ class SupervisorData(object):
                 return hs
 
     @property
-    def serverurl(self) -> str:
+    def server_url(self) -> str:
         return self.supervisord.options.serverurl
 
     @property
-    def serverport(self):
+    def server_port(self):
         return self.server_config['port']
 
     @property
@@ -188,7 +188,7 @@ class SupervisorData(object):
 
     def get_env(self) -> Dict[str, str]:
         """ Return a simple environment that can be used for the configuration of the XML-RPC client. """
-        return {'SUPERVISOR_SERVER_URL': self.serverurl,
+        return {'SUPERVISOR_SERVER_URL': self.server_url,
                 'SUPERVISOR_USERNAME': self.username,
                 'SUPERVISOR_PASSWORD': self.password}
 
@@ -241,8 +241,8 @@ class SupervisorData(object):
             self.logger.debug('SupervisorData.replace_tail_handlers: Server running without any HTTP'
                               ' authentication checking')
         # replace Supervisor handlers considering the order in supervisor.http.make_http_servers
-        self.httpserver.handlers[1] = tail_handler
-        self.httpserver.handlers[2] = main_tail_handler
+        self.http_server.handlers[1] = tail_handler
+        self.http_server.handlers[2] = main_tail_handler
 
     def replace_default_handler(self) -> None:
         """ This method replaces Supervisor web UI with Supvisors web UI. """
@@ -260,7 +260,7 @@ class SupervisorData(object):
             self.logger.debug('SupervisorData.replace_default_handler: Server running without any HTTP'
                               ' authentication checking')
         # replace Supervisor default handler at the end of the list
-        self.httpserver.handlers[-1] = def_handler
+        self.http_server.handlers[-1] = def_handler
 
     def close_httpservers(self) -> None:
         """ Call the close_httpservers of Supervisor.