+++ supvisors/tests/test_supervisordata.py
@@ -68,11 +68,11 @@ def test_creation(supervisor, source):
 def test_accessors(source):
     """ Test the accessors. """
     # test consistence with DummySupervisor configuration
-    assert source.httpserver is source.supervisord.options.httpserver
+    assert source.http_server is source.supervisord.options.httpserver
     assert source.supervisor_rpc_interface.rpc_name == 'supervisor_RPC'
     assert source.supvisors_rpc_interface.rpc_name == 'supvisors_RPC'
-    assert source.serverurl == f'http://{gethostname()}:65000'
-    assert source.serverport == 65000
+    assert source.server_url == f'http://{gethostname()}:65000'
+    assert source.server_port == 65000
     assert source.username == 'user'
     assert source.password == 'p@$$w0rd'
     assert source.supervisor_state == SupervisorStates.RUNNING